test/story_generator/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
Flask==3.0.2
Flask-WTF==1.2.1 